Backend 3

+++ application/views/funcionario/index.php
@@ -42,27 +42,28 @@
             </thead>
 
             <tbody>
-              <?php foreach ($funcionarios as $funcionario): ?>
-                <tr>
-                  <td class="text-center"><?= $funcionario->nome; ?></td>
-                  <td class="text-center"><?= $funcionario->email; ?></td>
-                  <td class="text-center">
-                    <?php echo ($funcionario->sexo == '0')? "Masculino" : "Feminino"; ?>
-                  </td>
-                  <td class="text-center">
-                    <?php echo $funcionario->data_nascimento; ?>
-                  </td>
-                  <td class="text-center">
-                    <a title="Editar funcionário" href="<?= site_url('funcionario/editar/'.$funcionario->id_funcionario)?>" class="btn btn-primary">
-                      <span class="fa fa-edit"></span></a>
+              <?php if (!is_null($funcionarios)): ?>
+                  <?php foreach ($funcionarios as $funcionario): ?>
+                    <tr>
+                      <td class="text-center"><?= $funcionario->nome; ?></td>
+                      <td class="text-center"><?= $funcionario->email; ?></td>
+                      <td class="text-center">
+                        <?php echo ($funcionario->sexo == '0')? "Masculino" : "Feminino"; ?>
+                      </td>
+                      <td class="text-center">
+                        <?php echo $funcionario->data_nascimento; ?>
+                      </td>
+                      <td class="text-center">
+                        <a title="Editar funcionário" href="<?= site_url('funcionario/editar/'.$funcionario->id_funcionario)?>" class="btn btn-primary">
+                        <span class="fa fa-edit"></span></a>
                         <button data-href="funcionario/excluir/<?php echo $funcionario->id_funcionario?>" class="btn btn-danger" title="Excluir funcionário" data-toggle="modal" data-target="#modalRemover">
                           <span class="fa fa-times"></span>
                         </button>
-                      </td>
-                    </tr>
-                  <?php endforeach ?>
+                       </td>
+                     </tr>
+                   <?php endforeach ?>
+                  <?php endif;?>
                 </tbody>
-
               </table>
             </div>
           </div>
